Make the mite file name code real simple.
This code has had lots of bugs and it has to be coordinated between
two modules, so make it simple.

Leave their filename intact, just take our extensions on the end.
.mite to indicate what it is and .pmc so Perl will pick it up.

Chose .pmc because that's what it is, a compiled Perl module, and
that's just what .pmc is for.  But .pmc's replace, not add to,
.pm files when loading, so we still need a different filename.